ensure consistent replica count for indexed-search
I got feedback from a customer that they used the resource estimator and
it gave back different replica counts for zoekt-indexserver and
zoekt-webserver. This confused him and sent him down a rabbithole which
wasted his time.

This adds an invariant which sets the replica count between the two to
always be the max between them.

Test Plan: Added a test which ensures replica counts are the same. It
failed before the code change.