add golden tests
It is hard to understand the effects of the resource estimator right
now. This adds some golden tests so as you make changes you can see the
effect by looking at diffs and/or diffing between different inputs.

I've just tested the default settings, as well as taking the defaults
but making it for a monorepo.

Test Plan: go test